+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `14:30` -- **Updated** `apps/backend/supabase/functions/_shared/customJobsParser.ts`
 - `14:29` -- **Updated** `apps/backend/supabase/functions/_shared/customJobsParser.ts`
 - `14:28` -- **Updated** `apps/backend/supabase/functions/_shared/advancedMatching.ts`
 - `14:27` -- **Updated** `apps/backend/supabase/functions/_shared/advancedMatching.ts`